Update ping check
